Rename "imitate" command category to "text"
